+++ Demo/Parallel.build.ps1
@@ -7,7 +7,7 @@
 	Invoke-Build * Parallel.build.ps1
 #>
 
-. .\SharedScript.ps1
+. .\Shared.ps1
 
 # The build engine defines aliases Invoke-Build and Invoke-Builds
 task Alias {
@@ -43,8 +43,8 @@ Four parallel builds of this test could be invoked as simple as this
 
 	Invoke-Builds @(
 		@{File='Dynamic.build.ps1'}
-		@{File='ErrorCases.build.ps1'}
-		@{File='Dynamic.build.ps1'; Task='Task1'}
+		@{File='Protected.build.ps1'; Task='Error1'}
+		@{File='Dynamic.build.ps1'; Task='Task0'}
 		@{File='Conditional.build.ps1'; Parameters=@{Configuration='Debug'}}
 	)
 
@@ -55,7 +55,7 @@ techniques useful for advanced build result analysis.
 task Many {
 	# These build parameter hashes are not changed (to be tested).
 	$build0 = @{File='Dynamic.build.ps1'}
-	$build1 = @{File='ErrorCases.build.ps1'; Task='Error1'}
+	$build1 = @{File='Protected.build.ps1'; Task='Error1'}
 	$build2 = @{File='Dynamic.build.ps1'; Task='Task0'}
 	$build3 = @{File='Conditional.build.ps1'; Parameters=@{Configuration='Debug'}}
 
@@ -89,7 +89,7 @@ task Many {
 	assert ($build[0].Result.Value)
 
 	# The call itself failed.
-	assert ($message -like "Failed builds:*Build: *\ErrorCases.build.ps1*ERROR: Error1*At *\SharedTasksData.tasks.ps1:*")
+	assert ($message -like "Failed builds:*Build: *\Protected.build.ps1*ERROR: Error1*At *\Protected.build.ps1:*")
 
 	# Check each build error; note that three builds succeeded because the
 	# parallel build engine lets all builds to work, even if some fail.